Added basic JS test used to ID classes and modify content
This is basically a test for my own knowledge at this
point.  Pingback.js runs through the document
looking for specific classes, then replaces the list
element with data stripped out of the UL's ID.

This is the basic method I think I will initially
implement.  It is simple enough to get something
working, and can be revisited later.